# swagger.yaml
openapi: 3.0.0
info:
  title: Legal Advice
  contact: {}
  version: '1.0'
servers:
- url: https://legal-advice-1812.onrender.com/api/v1
  variables: {}
- url: http://localhost:1812/api/v1
  variables: {}
paths:
...
  /questions:
    get:
      tags:
      - Question
      summary: Get All Questions
      operationId: GetAllQuestions
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - Question
      summary: Add Question
      operationId: AddQuestion
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              - title
              - subTitle[0][question]
              - subTitle[0][answer]
              - subTitle[1][question]
              - subTitle[1][answer]
              type: object
              properties:
                image:
                  type: string
                  format: binary
                title:
                  type: string
                  example: القانون الجنائي
                subTitle[0][question]:
                  type: string
                  example: سوال 6
                subTitle[0][answer]:
                  type: string
                  example: اجابة 6
                subTitle[1][question]:
                  type: string
                  example: سوال 66
                subTitle[1][answer]:
                  type: string
                  example: اجابة 66
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /questions/{id}:
    get:
      tags:
      - Question
      summary: Get Specific Question
      operationId: GetSpecificQuestion
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    patch:
      tags:
      - Question
      summary: Update Specific Question
      operationId: UpdateSpecificQuestion
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              - title
              type: object
              properties:
                image:
                  type: string
                  format: binary
                title:
                  type: string
                  example: الاوراق المالية
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Question
      summary: Remove Specific Question
      operationId: RemoveSpecificQuestion
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /posts:
    post:
      tags:
      - Post
      summary: Add Post
      operationId: AddPost
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Post
      summary: Get All Posts
      operationId: GetAllPosts
      parameters:
      - name: limit
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 5
      - name: page
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 2
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /posts/{id}:
    patch:
      tags:
      - Post
      summary: Update Post
      operationId: UpdatePost
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              - text
              type: object
              properties:
                image:
                  type: string
                  format: binary
                text:
                  type: string
                  example: Test Up
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Post
      summary: Get Specific Posts
      operationId: GetSpecificPosts
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Post
      summary: Delete Specific Posts
      operationId: DeleteSpecificPosts
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /posts/{id}/likes:
    post:
      tags:
      - Post
      summary: Add Like
      operationId: AddLike
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Post
      summary: Get All Likes
      operationId: GetAllLikes
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Post
      summary: Delete Like
      operationId: DeleteLike
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...